# Use call overload in FFTW.plan_*