--- a/base/fftw.jl
+++ b/base/fftw.jl
@@ -343,18 +343,25 @@ end
 
 call(w::ScaleWrap, Z) = scale!(w.w(Z), w.nrm)
 
-immutable PlanWrapR{Ti, To, Nd, P}
+immutable PlanWrapR{Ti, To, Copy, Nd, P}
     p::P
     osize::NTuple{Nd, Int}
 end
 
-call{Ti, To, Nd, P}(::Type{PlanWrapR{Ti, To, Nd}}, p::P, osize) =
-    PlanWrapR{Ti, To, Nd, P}(p, (osize...))
+call{Ti, To, Copy, Nd, P}(::Type{PlanWrapR{Ti, To, Copy, Nd}}, p::P, osize) =
+    PlanWrapR{Ti, To, Copy, Nd, P}(p, (osize...))
 
-function call{Ti, To}(w::PlanWrapR{Ti, To}, Z::StridedArray{Ti})
+function call{Ti, To}(w::PlanWrapR{Ti, To, false}, Z::StridedArray{Ti})
     assert_applicable(w.p, Z)
     W = Array(To, w.osize...)
     execute(w.p.plan, Z, W)
+    return W
+end
+
+function call{Ti, To}(w::PlanWrapR{Ti, To, true}, Z::StridedArray{Ti})
+    assert_applicable(w.p, Z)
+    W = Array(To, w.osize...)
+    execute(w.p.plan, copy(Z), W)
     return W
 end
 
@@ -603,7 +610,7 @@ for (Tr,Tc) in ((:Float32,:Complex64),(:Float64,:Complex128))
             osize[d1] = osize[d1]>>1 + 1
             Y = Array($Tc, osize...)
             p = Plan(X, Y, region, flags, tlim)
-            return PlanWrapR{$Tr, $Tc, ndims(X)}(p, osize)
+            return PlanWrapR{$Tr, $Tc, false, ndims(X)}(p, osize)
         end
 
         # FFTW currently doesn't support PRESERVE_INPUT for
@@ -654,7 +661,7 @@ for (Tr,Tc) in ((:Float32,:Complex64),(:Float64,:Complex128))
             osize[region] = d
             Y = Array($Tr, osize...)
             p = Plan(X, Y, region, flags | PRESERVE_INPUT, tlim)
-            return PlanWrapR{$Tc, $Tr, ndims(X)}(p, osize)
+            return PlanWrapR{$Tc, $Tr, false, ndims(X)}(p, osize)
         end
 
         function plan_brfft(X::StridedArray{$Tc}, d::Integer, region,
@@ -669,13 +676,7 @@ for (Tr,Tc) in ((:Float32,:Complex64),(:Float64,:Complex128))
             Y = Array($Tr, osize...)
             X = copy(X)
             p = Plan(X, Y, region, flags, tlim)
-            return Z::StridedArray{$Tc} -> begin
-                assert_applicable(p, Z)
-                Z = copy(Z)
-                W = Array($Tr, osize...)
-                execute(p.plan, Z, W)
-                return W
-            end
+            return PlanWrapR{$Tc, $Tr, true, ndims(X)}(p, osize)
         end
 
         rfft(X::StridedArray) = rfft(X, 1:ndims(X))