I'm aware of https://github.com/JuliaLang/julia/pull/6193. However, I think it is still useful to get this into 0.4 for following reasons.

1. This PR does not break the API

2. The API in the PR does not conflict with the one in #6193 (the commits themselves might)

3. Unless we're going to fix the anonymous function performance issue, this PR has a non-negligible performance improvment.

    I have a time propagator which uses FFT to transform between position and momentum space and this PR boost the performance by 23% (2.6s to 2.0s) and reduce the allocation number by 10x and size by 130M (2720 k -> 300k times, 440M -> 310M). These are numbers that includes the time spend for other calculations. (cleaner benchmark: https://github.com/JuliaLang/julia/pull/11994#issuecomment-118193390)

A few other notes about this PR:

1. I breaks the PR into several commits mainly because the nature of this PR means there's a lot of code movement and having separate commits makes it easier to write, test and review.

2. Is [this](https://github.com/JuliaLang/julia/blob/8ad0c1ca37271bdae8e4ad209774b3489cef770d/base/fftw.jl#L768) missing a `assert_applicable`?

@stevengj