@yuyichao 2015-07-02
Also worth noting that the performance improvement is 2-3x the one I get by specifying `FFTW.MEASURE`. (2.15s -> 2.0s)