@yuyichao 2015-07-02
Better benchmark. As expected, the speed up is the most visible with small input size (in my case the size is ~ `1000`)

Tested with
```julia
function test_fft(ary)
    println(length(ary))
    plan = plan_fft!(copy(ary), 1:1, FFTW.MEASURE)
    iplan = plan_ifft!(copy(ary), 1:1, FFTW.MEASURE)
    plan(ary)
    gc()
    @time for i in 1:10_000
        plan(ary)
        iplan(ary)
    end
end

for s in (10, 100, 10000, 16, 128, 16384)
    test_fft(rand(Complex{Float64}, s))
end
```

Current master
```
10
  74.272 milliseconds (351 k allocations: 17652 KB)
100
  75.764 milliseconds (340 k allocations: 17188 KB)
10000
   1.572 seconds      (340 k allocations: 17188 KB)
16
  60.032 milliseconds (340 k allocations: 17188 KB)
128
  75.934 milliseconds (340 k allocations: 17188 KB)
16384
   2.661 seconds      (340 k allocations: 17188 KB)
```

This PR
```
10
  12.091 milliseconds (61495 allocations: 1250 KB)
100
   9.360 milliseconds (50000 allocations: 781 KB)
10000
   1.470 seconds      (50000 allocations: 781 KB)
16
   3.248 milliseconds (50000 allocations: 781 KB)
128
  11.136 milliseconds (50000 allocations: 781 KB)
16384
   2.485 seconds      (50000 allocations: 781 KB)
```