@stevengj 2015-07-03
This will make #6193 significantly harder to merge because it will introduce lots of conflicts...