@ScottPJones 2015-07-03
I'd vote (if I had a vote!) for getting #6193 in first, and not have @stevengj spending time rebasing.  He's got much more important things to spend time on, IMO.