@yuyichao 2015-07-03
@stevengj An easy (although possibly stupid) way would be to just revert this in #6193.

I would definitely prefer having #6193 if it was not in the `0.5` milestone. This is by all mean a temporary solution and I'm only bringing it up because it gives a significant speed up for small and medium size arrays.

I didn't read #6193 very carefully. Is the structure/main part of the current `fftw.jl` still there? If the `plan_*` functions are just changed to return a different type than I'll be happy to match the name in it. Otherwise, I would argue that getting this performance improvement in 0.4 with this PR and revert it as part of #6193 is the way to go.

P.S. can you also have a look at the place that I suspect missing a assert?