@ViralBShah 2015-07-03
@ScottPJones I am not sure what you are voting on here and why you have a viewpoint on one developer's time is more important than another. It would be nice to restrict commenting where you are making a contribution or raising an issue. It is ok to upvote once in a while, and only if you are part of the implementation. I realize that you are excited and want to chime in on everything - but please think before chiming in, as it could end up being noise rather than adding value.