@yuyichao 2015-07-03
Actually this shouldn't introduce any bad conflict since the [first commit](https://github.com/stevengj/julia/commit/7fd42983b999fd0ef2a9b902976656e97492e237) in #6193 removes (move/split) this file and the conflict resolution should just be a reset.

It would indeed be better if this is changed to a non-breaking and no-function-added part of #6193 (basically moving code and returning functors in `plan_*`) but that might be too much work. I'm willing to help/do the splitting it if this plan sounds better though.

Other than that, I still believe it is nice to have this performance gain in for 0.4.