@ViralBShah 2015-07-03
Are we likely to merge #6193 for 0.4? If not, it seems worthwhile to at least have these gains on 0.4, perhaps after we branch, and target #6193 for 0.5. Also, we could revert this when we want to merge #6193 for 0.5. It appears that this is a very nice performance gain and nice to have.