@yuyichao 2015-07-03
> Are we likely to merge #6193 for 0.4?

It's on 0.5 milestone so my impression is unlikely.

> If not, it seems worthwhile to at least have these gains on 0.4, perhaps after we branch, and target #6193 for 0.5. Also, we could revert this when we want to merge #6193 for 0.5. It appears that this is a very nice performance gain and nice to have.

Yep. That's exactly my plan. It shouldn't be too bad and I'm waiting to see if there's any change I change in this PR to make @stevengj's work on #6193 easier.