@stevengj 2015-07-08
I'd rather just rip out the pure-Julia FFT from 6193.  That should be mainly a simple matter of just deleting a couple of files from that patch.