@yuyichao 2015-07-09
Close in favor of #12087 .